Clean up init files with only a comment, move the contents from api/__init__.py to api/app.py for cleaner imports
