--- rag-app/backend/__init__.py
@@ -1,3 +1,0 @@
-"""
-Backend package initialization
-"""
