--- rag-app/backend/auth/__init__.py
@@ -1,4 +1,0 @@
-"""
-Authentication and authorization package
-Simple demo implementation with user switching
-"""
